pyproject.toml
[project]
name = "flood-area-segmentation"
version = "0.1.0"
description = "Segmentation of flooded areas on aerial images using deep learning"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "albumentations>=2.0.8",
    "dvc[s3]>=3.65.0",
    "fire>=0.7.1",
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.8",
    "mlflow>=3.8.0",
    "numpy>=2.4.0",
    "onnxruntime-gpu>=1.23.2",
    "onnxscript>=0.5.7",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "pytorch-lightning>=2.6.0",
    "requests>=2.32.5",
    "segmentation-models-pytorch>=0.5.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]

[project.scripts]
flood-seg = "flood_area_segmentation.commands:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP"]
ignore = []

[tool.ruff.format]
quote-style = "double"

[dependency-groups]
dev = [
    "onnx>=1.20.0",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
